project/requirements.txt
tensorflow
mysqlclient
numpy
Pillow